apps/api/pyproject.toml
[project]
description = "Ashyq Bilim Backend"
name = "ashyq-bilim-api"
version = "0.1.0"
requires-python = ">=3.14.2,<3.15.0"
dependencies = [
  "fastapi[standard-no-fastapi-cloud-cli]>=0.135.1",
  "langchain>=1.2.12",
  "langchain-chroma>=1.1.0",
  "langchain-community>=0.4.1",
  "langchain-core>=1.2.19",
  "langchain-openai>=1.1.11",
  "langchain-text-splitters>=1.1.1",
  "openai>=2.28.0",
  "psycopg[binary,pool]>=3.3.3",
  "pydantic[email]>=2.12.5",
  "redis>=7.3.0",
  "resend>=2.24.0",
  "sqlmodel>=0.0.37",
  "typer>=0.24.1",
  "chromadb>=1.5.5",
  "alembic>=1.18.4",
  "alembic-postgresql-enum>=1.10.0",
  "stripe>=14.4.1",
  "logfire[fastapi,sqlalchemy]>=4.29.0",
  "python-ulid>=3.1.0",
  "argon2-cffi>=25.1.0",
  "orjson>=3.11.7",
  "cachetools>=7.0.5",
  "slowapi>=0.1.9",
  "pyjwt>=2.12.1",
  "langgraph>=1.1.2",
  "pydantic-settings>=2.13.1",
  "selectolax>=0.4.7",
]

[dependency-groups]

dev = [
  "faker>=40.11.0",
  "pytest>=9.0.2",
  "pytest-asyncio>=1.3.0",
  "pytest-cov>=7.0.0",
  "ruff>=0.15.6",
]

[tool.ruff]
target-version = "py315"
# lint.select = ["ALL"]
lint.fixable = ["ALL"]
lint.ignore = [
  "F401",
  "COM812",  # missing trailing comma (stylistic, optional)
  "DTZ005",  # datetime.now() used without tzinfo (not always relevant)
  "T20",     # flake8-print (allow print statements for debugging/logging)
  "TC002",   # type checking import block placement
  "TC001",   # type checking import not at top of file
  "TC003",   # type checking import formatting
  "PLR2004", # magic value used in comparison (often acceptable)
  "D",       # all docstring-related rules disabled
  "D100",    # missing docstring in public module
  "S101",    # Use of `assert` detected
  "TD002",   # Missing author in TODO;
  "S106",
  "TD003",
  "S105",
  "ARG002",
  "ANN001",
  "S105",
  "PLR0913",
  "ANN201",
  "PLC0415",
  "TRY003",
  "EM101",
]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = ["asyncio: mark test as async"]

[tool.logfire]
ignore_no_config = true

[tool.pyrefly]
project-includes = ["**/*.py*"]
